+++ compiler/src/main/java/info/teksol/mc/messages/SourcePositionTranslator.java
@@ -5,6 +5,7 @@ import org.jspecify.annotations.NullMarked;
 
 @NullMarked
 public interface SourcePositionTranslator {
+    SourcePositionTranslator EMPTY = position -> position;
 
     SourcePosition apply(SourcePosition position);
 }